Fixing some bugs from the refactor. Preserving component state when navigating away. Better comments and organization.
