+++ rlbot_gui/match_runner/match_runner.py
@@ -4,8 +4,10 @@ from rlbot.matchconfig.match_config import PlayerConfig, MatchConfig, MutatorCon
 from rlbot.parsing.incrementing_integer import IncrementingInteger
 from rlbot.setup_manager import SetupManager
 from rlbot.utils.structures.bot_input_struct import PlayerInput
-from rlbot.utils.game_state_util import GameState, CarState, Physics, Vector3, Rotator, BallState, GameInfoState
+from rlbot.utils.game_state_util import GameState, CarState, Physics, Vector3, Rotator
 from rlbot.utils.structures.game_data_struct import GameTickPacket
+
+from rlbot_gui.type_translation.set_state_translation import dict_to_game_state
 
 sm: SetupManager = None
 
@@ -61,6 +63,7 @@ def spawn_car_in_showroom(loadout_config: LoadoutConfig, team: int):
         rotation=Rotator(0, 0, 0)
     ))}))
 
+
 def set_game_state(state):
     global sm
     if sm is None:
@@ -69,56 +72,6 @@ def set_game_state(state):
     game_state = dict_to_game_state(state)
     sm.game_interface.set_game_state(game_state)
 
-def dict_to_game_state(state_dict):
-    gs = GameState()
-    if 'ball' in state_dict:
-        gs.ball = BallState()
-        if 'physics' in state_dict['ball']:
-            gs.ball.physics = dict_to_physics(state_dict['ball']['physics'])
-    if 'cars' in state_dict:
-        gs.cars = {}
-        for index, car in state_dict['cars'].items():
-            car_state = CarState()
-            if 'physics' in car:
-                car_state.physics = dict_to_physics(car['physics'])
-            gs.cars[int(index)] = car_state
-    if 'game_info' in state_dict:
-        gs.game_info = GameInfoState()
-        if 'paused' in state_dict['game_info']:
-            gs.game_info.paused = state_dict['game_info']['paused']
-    return gs
-
-def dict_to_physics(physics_dict):
-    phys = Physics()
-    if 'location' in physics_dict:
-        phys.location = dict_to_vec(physics_dict['location'])
-    if 'velocity' in physics_dict:
-        phys.velocity = dict_to_vec(physics_dict['velocity'])
-    if 'angular_velocity' in physics_dict:
-        phys.angular_velocity = dict_to_vec(physics_dict['angular_velocity'])
-    if 'rotation' in physics_dict:
-        phys.rotation = dict_to_rot(physics_dict['rotation'])
-    return phys
-
-def dict_to_vec(v):
-    vec = Vector3()
-    if 'x' in v:
-        vec.x = v['x']
-    if 'y' in v:
-        vec.y = v['y']
-    if 'z' in v:
-        vec.z = v['z']
-    return vec
-
-def dict_to_rot(r):
-    rot = Rotator()
-    if 'pitch' in r:
-        rot.pitch = r['pitch']
-    if 'yaw' in r:
-        rot.yaw = r['yaw']
-    if 'roll' in r:
-        rot.roll = r['roll']
-    return rot
 
 def fetch_game_tick_packet() -> GameTickPacket:
     global sm
@@ -128,6 +81,7 @@ def fetch_game_tick_packet() -> GameTickPacket:
     game_tick_packet = GameTickPacket()
     sm.game_interface.update_live_data_packet(game_tick_packet)
     return game_tick_packet
+
 
 def start_match_helper(bot_list, match_settings):
     print(bot_list)
@@ -196,4 +150,3 @@ def shut_down():
         sm.shut_down(time_limit=5, kill_all_pids=True)
     else:
         print("There gotta be some setup manager already")
-
